refactor: split hercule-resolve-http off from the main package
