+++ packages/hercule-core/package.json
@@ -31,7 +31,6 @@
     "clone-regexp": "^1.0.0",
     "duplexer3": "^0.1.4",
     "get-stream": "^3.0.0",
-    "got": "^8.0.0",
     "isstream": "^0.1.2",
     "left-split": "^1.0.0",
     "lodash": "^4.0.0",
